2pc - presumed commit: increment counter of ok prepared subtxns
To implement presumed commit optimization we track the
number of successfully prepared transactions.

If any prepare fails, we must send ROLLBACK messages to already
prepared transactions and must not delete the record from the
'_transaction' space until all already prepared transaction aborted.

When a prepared transaction aborted, it must send ACK message to
the coordinator. This ACK message decrements the counter in the
'_transaction' space record.  When the counter became zero, we sure
that all prepared transaction were aborted and now can delete the
record from the '_transaction'.

This patch only introduces the counter.