vinyl: start scheduler for remote recovery
The amount of data sent when bootstrapping replica is only limited by
the master's disk size, which can exceed the size of available memory by
orders of magnitude. So we need the scheduler to be up and running while
bootstrapping from the remote host so that it could schedule dumps when
the quota limit is hit.