vinyl: use vy_task.run_count to complete compact task
When we will have single mem per index, there will be cases,
when compaction of the index runs and compaction of a range are
working concurrently.
Index compaction 'complete' callback can update compact_priority before
ending of the range compaction, and vy_task_compact_complete() will
delete the new run due to the compact_priority.

So we must save the compact_priority in the compaction task sheduling.